requirements.txt
requests
pyyaml
prettytable
clint
toml